pyproject.toml
[project]
name = "fastaudio"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "dishka>=1.6.0",
    "fastapi>=0.116.1",
    "greenlet>=3.2.3",
    "mypy>=1.17.0",
    "pydantic-settings>=2.10.1",
    "ruff>=0.12.7",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
]
